// frontend/src/client/models/NewIngredient.ts
// tslint:disable
/**
 * Macro Macro API
 * An API that provides access to recipe and ingredient information, both user-generated and sourced from the USDA
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Macros,
    MacrosFromJSON,
    MacrosFromJSONTyped,
    MacrosToJSON,
} from './';

/**
 * 
 * @export
 * @interface NewIngredient
 */
export interface NewIngredient extends Macros {
    /**
     * 
     * @type {string}
     * @memberof NewIngredient
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof NewIngredient
     */
    amount: number;
    /**
     * 
     * @type {string}
     * @memberof NewIngredient
     */
    unit: string;
}

export function NewIngredientFromJSON(json: any): NewIngredient {
    return NewIngredientFromJSONTyped(json, false);
}

export function NewIngredientFromJSONTyped(json: any, ignoreDiscriminator: boolean): NewIngredient {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        ...MacrosFromJSONTyped(json, ignoreDiscriminator),
        'name': json['name'],
        'amount': json['amount'],
        'unit': json['unit'],
    };
}

export function NewIngredientToJSON(value?: NewIngredient): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        ...MacrosToJSON(value),
        'name': value.name,
        'amount': value.amount,
        'unit': value.unit,
    };
}